+++ cmd/apictl-key/main.go
@@ -1,16 +1,18 @@
 package main
 
 import (
+	"os"
+
 	"github.com/spf13/cobra"
 )
-
-var apictl_key = &cobra.Command{
-	Use: "apictl-key [command]",
-}
 
 // Version is inserted at build using --ldflags -X
 var Version = "(unknown version)"
 
+var argparser = &cobra.Command{
+	Use: os.Args[0],
+}
+
 func main() {
-	apictl_key.Execute()
+	argparser.Execute()
 }