+++ tests/cluster/external_test.go
@@ -9,6 +9,7 @@ import (
 	"io/ioutil"
 	"net/http"
 	"net/url"
+	"os/exec"
 	"strings"
 	"testing"
 
@@ -104,6 +105,9 @@ func inArray(needle string, haystack []string) bool {
 }
 
 func TestHTTPExternalModify(t *testing.T) {
+	if _, err := exec.LookPath("docker"); err != nil {
+		t.Skip("docker not installed")
+	}
 	assert := &testutil.Assert{T: t}
 	res, body := doRequest(t, "https://ambassador.standalone.svc.cluster.local/external-http/headers") //nolint:bodyclose
 	// HTTP/1.1 200 OK
@@ -150,6 +154,9 @@ func TestHTTPExternalModify(t *testing.T) {
 }
 
 func TestHTTPExternalIntercept(t *testing.T) {
+	if _, err := exec.LookPath("docker"); err != nil {
+		t.Skip("docker not installed")
+	}
 	assert := &testutil.Assert{T: t}
 	res, body := doRequest(t, "https://ambassador.standalone.svc.cluster.local/external-http/ip") //nolint:bodyclose
 	// HTTP/1.1 404 Not Found
@@ -172,6 +179,9 @@ func TestHTTPExternalIntercept(t *testing.T) {
 }
 
 func TestGRPCExternalModify(t *testing.T) {
+	if _, err := exec.LookPath("docker"); err != nil {
+		t.Skip("docker not installed")
+	}
 	assert := &testutil.Assert{T: t}
 	res, body := doRequest(t, "https://ambassador.standalone.svc.cluster.local/external-grpc/headers") //nolint:bodyclose
 	// HTTP/1.1 200 OK
@@ -215,6 +225,9 @@ func TestGRPCExternalModify(t *testing.T) {
 }
 
 func TestGRPCExternalIntercept(t *testing.T) {
+	if _, err := exec.LookPath("docker"); err != nil {
+		t.Skip("docker not installed")
+	}
 	assert := &testutil.Assert{T: t}
 	res, body := doRequest(t, "https://ambassador.standalone.svc.cluster.local/external-grpc/ip") //nolint:bodyclose
 	// HTTP/1.1 200 OK